refactor kv benchmark under cli-bench
