+++ src/cli-bench/src/kv/get.rs
@@ -17,17 +17,17 @@ use std::{sync::Arc, time::Duration};
 use clap::Parser;
 use dashmap::DashMap;
 use futures::future;
-use prettytable::{row, Table};
-use protocol::placement_center::placement_center_kv::{GetRequest, SetRequest};
-use rand::{thread_rng, Rng};
-use serde::{de::DeserializeOwned, Serialize};
 use tokio::time::Instant;
 
-use crate::{error::BenchMarkError, BenchMark};
-use grpc_clients::{
-    placement::kv::call::{placement_get, placement_set},
-    pool::ClientPool,
+use crate::{
+    error::BenchMarkError,
+    kv::utils::{
+        do_placement_get, do_placement_set, gen_data, print_latency_table, print_throughput_table,
+        print_total_statistics, sort_lantencies_by_client_id,
+    },
+    BenchMark,
 };
+use grpc_clients::pool::ClientPool;
 
 #[derive(Debug, Clone, Parser)]
 pub struct KvGetBenchArgs {
@@ -40,7 +40,7 @@ pub struct KvGetBenchArgs {
     pub worker_threads: usize,
 
     /// The number of keys to get
-    #[clap(long, default_value = "10000")]
+    #[clap(long, default_value = "500")]
     pub num_keys: usize,
 
     /// The size of each value in bytes
@@ -50,56 +50,6 @@ pub struct KvGetBenchArgs {
     /// The address of the placement center
     #[clap(long, default_value = "127.0.0.1:1228", value_delimiter = ',')]
     pub pc_addrs: Vec<String>,
-}
-
-async fn do_placement_get<T>(
-    client_pool: &ClientPool,
-    addrs: &[&str],
-    req: GetRequest,
-) -> Result<T, BenchMarkError>
-where
-    T: DeserializeOwned,
-{
-    let reply = placement_get(client_pool, addrs, req)
-        .await
-        .map_err(Box::new)?;
-
-    let ret = serde_json::from_str(reply.value.as_str())?;
-
-    Ok(ret)
-}
-
-async fn do_placement_set<T>(
-    client_pool: &ClientPool,
-    addrs: &[&str],
-    key: String,
-    value: T,
-) -> Result<(), BenchMarkError>
-where
-    T: Serialize,
-{
-    let req = SetRequest {
-        key,
-        value: serde_json::to_string(&value)?,
-    };
-
-    placement_set(client_pool, addrs, req)
-        .await
-        .map_err(Box::new)?;
-
-    Ok(())
-}
-
-// generate random data of specified length, consisting of alphanumeric characters
-fn gen_data(length: usize) -> String {
-    let mut rng = thread_rng();
-    let chars: Vec<char> = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-        .chars()
-        .collect();
-
-    (0..length)
-        .map(|_| chars[rng.gen_range(0..chars.len())])
-        .collect()
 }
 
 #[axum::async_trait]
@@ -124,7 +74,7 @@ impl BenchMark for KvGetBenchArgs {
             num_clients, worker_threads, num_keys, value_size, pc_addrs
         );
 
-        let client_pool = Arc::new(ClientPool::new(num_clients as u64));
+        let client_pool = Arc::new(ClientPool::new(0)); // unlimited connections
         let mut set_handles = Vec::with_capacity(num_clients);
 
         for client_id in 0..num_clients {
@@ -188,8 +138,6 @@ impl BenchMark for KvGetBenchArgs {
 
                         total_key_size += key.len();
 
-                        let req = GetRequest { key };
-
                         let now = Instant::now();
 
                         // get the value back
@@ -199,7 +147,7 @@ impl BenchMark for KvGetBenchArgs {
                                 .iter()
                                 .map(String::as_str)
                                 .collect::<Vec<_>>(),
-                            req,
+                            key,
                         )
                         .await?;
 
@@ -225,93 +173,20 @@ impl BenchMark for KvGetBenchArgs {
 
         future::join_all(get_handles).await;
 
-        let total_time = total_now.elapsed();
-
-        println!(
-            "==================\nBenchmark KV get finished with:\n\t\ttime: {:?}s\n\t\tms/op: {}\n\t\top/ms: {}",
-            total_time.as_secs_f64(),
-            total_time.as_millis() as f64 / (num_clients * num_keys) as f64,
-            (num_clients * num_keys) as f64 / total_time.as_millis() as f64,
-        );
+        // print statistics across all clients
+        print_total_statistics(total_now.elapsed(), num_clients, num_keys);
 
         // Sort latencies by client ID
-        let mut latencies_sorted = latencies
-            .iter()
-            .map(|entry| (*entry.key(), entry.value().clone()))
-            .collect::<Vec<_>>();
-
-        latencies_sorted.sort_by_key(|(client_id, _)| *client_id);
+        let latencies_sorted = sort_lantencies_by_client_id(&latencies);
 
         // Print the results in a table format
-        let mut table_latency = Table::new();
-        let mut table_throughput = Table::new();
-
-        table_latency.add_row(row![
-            "Client ID",
-            "P50 Latency (us)",
-            "P90 Latency (us)",
-            "P99 Latency (us)",
-            "P999 Latency (us)",
-            "Average Latency (us)",
-            "Min Latency (us)",
-            "Max Latency (us)",
-            "Median Latency (us)"
-        ]);
-
-        table_throughput.add_row(row![
-            "Client ID",
-            "nMsg per second",
-            "Total message size (bytes)",
-            "Bytes per second"
-        ]);
-
-        for (client_id, latencies) in latencies_sorted.iter() {
-            let mut sorted_latencies = latencies.clone();
-            sorted_latencies.sort_unstable();
-
-            let p50 = sorted_latencies[sorted_latencies.len() / 2].as_micros();
-            let p90 = sorted_latencies[(sorted_latencies.len() * 9) / 10].as_micros();
-            let p99 = sorted_latencies[(sorted_latencies.len() * 99) / 100].as_micros();
-            let p999 = sorted_latencies[(sorted_latencies.len() * 999) / 1000].as_micros();
-            let average: u128 = sorted_latencies.iter().map(|d| d.as_micros()).sum::<u128>()
-                / sorted_latencies.len() as u128;
-            let min = sorted_latencies.first().unwrap().as_micros();
-            let max = sorted_latencies.last().unwrap().as_micros();
-            let median = sorted_latencies[sorted_latencies.len() / 2].as_micros();
-
-            table_latency.add_row(row![
-                client_id, p50, p90, p99, p999, average, min, max, median
-            ]);
-
-            let n_message = 2 * num_keys;
-
-            let total_time_secs = sorted_latencies
-                .iter()
-                .map(|d| d.as_secs_f64())
-                .sum::<f64>();
-
-            let n_message_per_second = n_message as f64 / total_time_secs;
-
-            let total_key_size = total_key_sizes
-                .get(client_id)
-                .map(|size| *size.value())
-                .expect("Client Id not found");
-
-            let total_value_size = num_keys * value_size;
-
-            let bytes_per_second = (total_key_size + total_value_size) as f64 / total_time_secs;
-
-            table_throughput.add_row(row![
-                client_id,
-                n_message_per_second,
-                total_key_size + total_value_size,
-                bytes_per_second
-            ]);
-        }
-
-        table_latency.printstd();
-
-        table_throughput.printstd();
+        print_latency_table(&latencies_sorted);
+        print_throughput_table(
+            &latencies_sorted,
+            num_keys,
+            value_size,
+            total_key_sizes.as_ref(),
+        );
 
         Ok(())
     }
